Refactor routes to be more modular
